+++ R/pip.R
@@ -68,6 +68,9 @@ pip <- function(country = "all",
   welfare_type <- match.arg(welfare_type)
   reporting_level <- match.arg(reporting_level)
   group_by <- match.arg(group_by)
+  if(length(lkup) <= 5)
+    stop("You are probably passing more than one dataset as lkup argument.
+  Try passing a single one by subsetting it lkup <- lkups$versions_paths$dataset_name_PROD")
   #Make country case insensitive
   country <- tolower(country)
   country <- if(country == "all") country else toupper(country)