+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## Enhancements
 - [Sort the rows by `country_code` and `reporting_year` in the pip response.](https://github.com/PIP-Technical-Team/pipapi/issues/248)
+- [Fix casing making pip call case insensitive](https://github.com/PIP-Technical-Team/pipapi/issues/120)
 
 
 # pipapi 1.0.0