+++ NEWS.md
@@ -3,6 +3,9 @@
 ## Enhancements
 - [Sort the rows by `country_code` and `reporting_year` in the pip response.](https://github.com/PIP-Technical-Team/pipapi/issues/248)
 
+## New features
+
+- [Add /valid_years as new endpoint to get the valid years information of the data](https://github.com/PIP-Technical-Team/pipapi/issues/182)
 
 # pipapi 1.0.0
 
@@ -13,7 +16,6 @@
 - Change parameter `version_length` for `vintage_pattern` in `extract_data_dirs()`. The algorithm for extracting valid versions has been modified. 
 - Add option to enable disk based caching of `pip()` and `pip_grp()`
 - Add asynchronous processing of slow API requests
-- [Add /valid_years as new endpoint to get the valid years information of the data](https://github.com/PIP-Technical-Team/pipapi/issues/182)
  
 ## Enhancements