+++ NEWS.md
@@ -2,6 +2,10 @@
 
 ## Enhancements
 - [Sort the rows by `country_code` and `reporting_year` in the pip response.](https://github.com/PIP-Technical-Team/pipapi/issues/248)
+- Region codes can now be passed directly to the `country` query parameter to 
+return all countries pertaining to the specified region
+- the `/valid-params` endpoint gains an `endpoint` parameter that allows to only
+return parameters that are relevant to the specified endpoint
 
 
 # pipapi 1.0.0