+++ NEWS.md
@@ -7,6 +7,9 @@ return all countries pertaining to the specified region
 - the `/valid-params` endpoint gains an `endpoint` parameter that allows to only
 return parameters that are relevant to the specified endpoint
 
+## New features
+
+- [Add /valid_years as new endpoint to get the valid years information of the data](https://github.com/PIP-Technical-Team/pipapi/issues/182)
 
 # pipapi 1.0.0
 
@@ -17,7 +20,6 @@ return parameters that are relevant to the specified endpoint
 - Change parameter `version_length` for `vintage_pattern` in `extract_data_dirs()`. The algorithm for extracting valid versions has been modified. 
 - Add option to enable disk based caching of `pip()` and `pip_grp()`
 - Add asynchronous processing of slow API requests
-- [Add /valid_years as new endpoint to get the valid years information of the data](https://github.com/PIP-Technical-Team/pipapi/issues/182)
  
 ## Enhancements